Make homepage editable by admins
Prior to this commit the home site was fully hardcoded, although
it's a good place to put some news there. There is a new settings table
in the db which where the "welcome message" is stored. It can be
formatted using markdown syntax.
Therefore it was necessary to require a lib for markdown through
composer.