UI overhaul
Added error message for when a user tries to see a non-existing
folder.

Minor changes to getParentPath and getDepth in Path class to better suit
the needs of the browser.

Download button is now disabled if the according file is not yet
approved.

Added notice to review notification saying it takes a while till
files are available throughout the whole network.

Changed dropzone.config.js to allow cancelling uploads and changed the
message inside the dropzone to be more clear. Also, a message is now
shown that the page needs to be refreshed, as soon as the first file
upload succeeds.

Added My Files button to the main menu to give it more presence
and especially be more clear to users. Removed the "go to my folder"
button from the browser.

Small design improvements to the browser.